requirements.txt
flake8==6.0.0
mypy==1.4.1
pandoc==2.3
PyYAML==6.0.1